+++ autocompletion.php
@@ -27,6 +27,8 @@ abstract class BaseApplication extends yii\base\Application
 /**
  * Class WebApplication
  * Include only Web application related components here
+ *
+ * @property User $user User component.
  */
 class WebApplication extends yii\web\Application
 {
@@ -37,5 +39,15 @@ class WebApplication extends yii\web\Application
  * Include only Console application related components here
  */
 class ConsoleApplication extends yii\console\Application
+{
+}
+
+/**
+ * User component
+ * Include only Web application related components here
+ *
+ * @property \common\models\User $identity User model.
+ */
+class User extends \yii\web\User
 {
 }