+++ backend/config/_base.php
@@ -2,11 +2,7 @@
 return [
     'id' => 'backend',
     'basePath' => dirname(__DIR__),
-    'homeUrl'=>Yii::getAlias('@backendUrl'),
     'components' => [
-        'urlManager'=>require(__DIR__.'/_urlManager.php'),
-        'errorHandler' => [
-            'errorAction' => 'site/error',
-        ],
+        'urlManager'=>require(__DIR__.'/_urlManager.php')
     ],
 ];