+++ common/components/behaviors/LocaleBehavior.php
@@ -40,9 +40,9 @@ class LocaleBehavior extends Behavior{
             $userLocale = Yii::$app->getRequest()->getCookies()->getValue($this->cookieName);
         } else {
             $userLocale = !Yii::$app->user->isGuest
-                        && Yii::$app->user->getIdentity()->profile
-                        && Yii::$app->user->getIdentity()->profile->locale
-                ? Yii::$app->user->getIdentity()->profile->locale
+                        && Yii::$app->user->getIdentity()->userProfile
+                        && Yii::$app->user->getIdentity()->userProfile->locale
+                ? Yii::$app->user->getIdentity()->userProfile->locale
                 : Yii::$app->request->getPreferredLanguage(array_keys(Yii::$app->params['availableLocales']));
         }
         Yii::$app->language = $userLocale;