+++ frontend/modules/user/controllers/DefaultController.php
@@ -11,18 +11,25 @@ use yii\web\Controller;
 
 class DefaultController extends Controller
 {
-    public function actions(){
+    /**
+     * @return array
+     */
+    public function actions()
+    {
         return [
             'avatar-upload'=>[
                 'class'=>UploadAction::className(),
-                'fileProcessing'=>function($file){
-                    Image::thumbnail($file->path->getPath(), 215,215)
+                'fileProcessing' => function($file) {
+                    Image::thumbnail($file->path->getPath(), 215, 215)
                         ->save($file->path->getPath());
                 }
             ]
         ];
     }
 
+    /**
+     * @return array
+     */
     public function behaviors()
     {
         return [
@@ -31,38 +38,44 @@ class DefaultController extends Controller
                 'rules' => [
                     [
                         'allow' => true,
-                        'roles' => ['@'],
-                    ],
-                ],
-            ],
+                        'roles' => ['@']
+                    ]
+                ]
+            ]
         ];
     }
 
+    /**
+     * @return string|\yii\web\Response
+     */
     public function actionIndex()
     {
         $user = Yii::$app->user->identity;
         $model = new AccountForm();
         $model->username = $user->username;
-        if($model->load($_POST) && $model->validate()){
+        if ($model->load($_POST) && $model->validate()) {
             $user->username = $model->username;
             $user->setPassword($model->password);
             $user->save();
             Yii::$app->session->setFlash('alert', [
-                'options'=>['class'=>'alert-success'],
-                'body'=>Yii::t('frontend', 'Your profile has been successfully saved')
+                'options' => ['class'=>'alert-success'],
+                'body' => Yii::t('frontend', 'Your account has been successfully saved')
             ]);
             return $this->refresh();
         }
         return $this->render('index', ['model'=>$model]);
     }
 
+    /**
+     * @return string|\yii\web\Response
+     */
     public function actionProfile()
     {
-        $model = Yii::$app->user->identity->profile;
-        if($model->load($_POST) && $model->save()){
+        $model = Yii::$app->user->identity->userProfile;
+        if ($model->load($_POST) && $model->save()) {
             Yii::$app->session->setFlash('alert', [
-                'options'=>['class'=>'alert-success'],
-                'body'=>Yii::t('frontend', 'Your profile has been successfully saved')
+                'options' => ['class'=>'alert-success'],
+                'body' => Yii::t('frontend', 'Your profile has been successfully saved', [], $model->locale)
             ]);
             return $this->refresh();
         }