+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Yii Framework 2 Change Log
 
 1.5.0 under development
 -----------------------
-...
+- Fix: FileCache now uses same path for all applications
 - Enh: common\components\behaviors\CacheInvalidateBehavior
 - ... Fixes ...