+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Yii Framework 2 Change Log
 
 1.5.0 under development
 -----------------------
-- Fix: FileCache now uses same path for all applications
+- Fix: Autocompletion support
+- Enh: FileCache now uses same path for all applications
 - Enh: common\components\behaviors\CacheInvalidateBehavior
 - ... Fixes ...