+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Yii Framework 2 Change Log
 
 1.5.0 under development
 -----------------------
+- Fix: contact form fix
+- Enh: "robot" email now ca be set in .env
 - Enh: #79 chosen locale is stored in cookies
 - Chg: application structure
 - Chg #59: dotenv support