+++ CHANGELOG.md
@@ -1,8 +1,12 @@
-Yii Framework 2 Change Log
+Yii Starter Kit Change Log
 ==========================
 
-1.5.0 under development
+2.0.0 under development
 -----------------------
+
+
+1.5.0
+-----
 - Enh: ``$cachingDuration`` parameter was added to ``common\components\keyStorage\KeyStorage::get``
 - Fix: contact form fix
 - Enh: "robot" email now ca be set in .env